Configure the layout model for full screen maps
