+++ nextjs/.env.example
@@ -1,1 +1,2 @@
 NEXT_PUBLIC_BACKEND_BASE_URL=http://localhost:8000
+NEXT_PUBLIC_MAPBOX_ACCESS_TOKEN=
