+++ nextjs/src/app/(app)/data-sources/inspect/[externalDataSourceId]/InspectExternalDataSource.tsx
