+++ nextjs/src/app/(app)/data-sources/inspect/[externalDataSourceId]/page.tsx
