+++ nextjs/src/app/(app)/login/login-form.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { gql, useMutation } from "@apollo/client";
-import { login } from "../../actions/auth";
+import { login } from "../../../actions/auth";
 import { useForm } from "react-hook-form";
 import { zodResolver } from "@hookform/resolvers/zod"
 import { z } from "zod";
