+++ nextjs/src/app/(app)/logout/page.tsx
@@ -1,5 +1,5 @@
 import { Metadata } from "next";
-import { useRequireAuth } from "../../hooks/auth";
+import { useRequireAuth } from "../../../hooks/auth";
 import LogoutForm from "./logout-form";
 
 // This has been split into a two components to separate the client-side (LogoutForm)
