+++ nextjs/src/components/navbar.tsx
@@ -17,27 +17,10 @@ interface NavbarProps {
 
 export default function Navbar({ isLoggedIn }: NavbarProps) {
   const pathname = usePathname()
-
-  function MappedIcon() {
-    return (
-      <svg xmlns="http://www.w3.org/2000/svg" width="29" height="43" viewBox="0 0 29 43" fill="none">
-        <circle cx="14.7351" cy="14.2833" r="13.9737" fill="#678DE3" />
-        <path d="M16.3481 39.9C15.6435 41.1625 13.8271 41.1625 13.1226 39.9L2.23064 20.3842C1.54355 19.153 2.43356 17.6371 3.84342 17.6371L25.6273 17.6371C27.0371 17.6371 27.9271 19.1531 27.24 20.3842L16.3481 39.9Z" fill="#678DE3" />
-      </svg>
-    )
-  }
-
-  function CollapsibleIcon() {
-    return (
-      <svg xmlns="http://www.w3.org/2000/svg" width="10" height="6" viewBox="0 0 10 6" fill="none">
-        <path d="M1.74599 0.12793L5.09712 3.47176L8.44825 0.12793L9.47769 1.15736L5.09712 5.53793L0.716553 1.15736L1.74599 0.12793Z" fill="white" />
-      </svg>
-    )
-  }
   return (
     <>
       {isLoggedIn ? (
-        <nav className='flex flex-row justify-start items-stretch gap-md font-IBMPlexSansCondensed text-lg border-b border-meepGray-700 px-sm'>
+        <nav className='sticky top-0 shrink-0 flex flex-row justify-start items-stretch gap-md font-IBMPlexSansCondensed text-lg border-b border-meepGray-700 px-sm'>
           <Link href='/' className="py-sm"><MappedIcon /></Link>
           <Link
             className={`link ${pathname === '/reports' ? 'active' : 'flex items-center'}`}
@@ -66,7 +49,7 @@ export default function Navbar({ isLoggedIn }: NavbarProps) {
       ) : (
         <nav className="p-sm">
           <ul className="flex flex-row">
-            <div className="flex flex-row justify-between items-center basis-1/2 bg-meepGray-700 rounded-lg">
+            <div className="shrink-0 flex flex-row justify-between items-center basis-1/2 bg-meepGray-700 rounded-lg">
               <div className="flex flex-row items-center gap-xs">
                 <Link className="pl-xs" href='/'><MappedIcon />
                 </Link>
@@ -110,3 +93,19 @@ export default function Navbar({ isLoggedIn }: NavbarProps) {
   )
 }
 
+function MappedIcon() {
+  return (
+    <svg xmlns="http://www.w3.org/2000/svg" width="29" height="43" viewBox="0 0 29 43" fill="none">
+      <circle cx="14.7351" cy="14.2833" r="13.9737" fill="#678DE3" />
+      <path d="M16.3481 39.9C15.6435 41.1625 13.8271 41.1625 13.1226 39.9L2.23064 20.3842C1.54355 19.153 2.43356 17.6371 3.84342 17.6371L25.6273 17.6371C27.0371 17.6371 27.9271 19.1531 27.24 20.3842L16.3481 39.9Z" fill="#678DE3" />
+    </svg>
+  )
+}
+
+function CollapsibleIcon() {
+  return (
+    <svg xmlns="http://www.w3.org/2000/svg" width="10" height="6" viewBox="0 0 10 6" fill="none">
+      <path d="M1.74599 0.12793L5.09712 3.47176L8.44825 0.12793L9.47769 1.15736L5.09712 5.53793L0.716553 1.15736L1.74599 0.12793Z" fill="white" />
+    </svg>
+  )
+}
